+++ CHANGELOG.md
@@ -29,6 +29,8 @@ This version adds the orders template and tests
 ### Changed
 - Added a selection field to the OrderDetail Schema in `src/api/v1/apps/orders/schemas.py`
 - Added services for items in `src/api/v1/apps/orders/services.py`
+- Added a service for change sended status in `src/api/v1/apps/orders/services.py`
+- Added the implementation to use the new method for change sended status in `src/api/v1/apps/orders/router.py`
 
 ### Fixed
 - N/A